@@ -1,39 +1,43 @@
 {
-  "name": "fretchen.github.io",
-  "homepage": "https://fretchen.github.io/",
-  "version": "0.1.0",
-  "description": "The personal website of fretchen",
   "scripts": {
-    "start": "vite",
-    "lint": "eslint src",
+    "dev": "vite",
+    "prebuild": "tsx ./utils/getBlogs.ts",
     "build": "vite build",
-    "prebuild": "ts-node src/utils/getBlogs.ts",
     "preview": "vite preview",
+    "lint": "eslint .",
+    "lint:fix": "eslint --fix .",
     "test": "mocha",
-    "typecheck": "tsc --noEmit",
-    "typewatch": "tsc --noEmit --watch"
+    "prepare": "panda codegen"
   },
-  "license": "MIT",
   "dependencies": {
+    "@mdx-js/mdx": "^3.1.0",
+    "@mdx-js/react": "^3.1.0",
+    "@mdx-js/rollup": "^3.1.0",
+    "@types/mocha": "^10.0.10",
+    "@vitejs/plugin-react": "^4.3.4",
+    "mocha": "^11.0.1",
     "mui-markdown": "^1.2.5",
-    "react": "^18.2.0",
-    "react-dom": "^18.2.0",
-    "ts-node": "^10.9.2"
+    "react": "^19.0.0",
+    "react-dom": "^19.0.0",
+    "tsx": "^4.19.2",
+    "vike-react": "^0.5.10"
   },
   "devDependencies": {
     "@eslint/js": "^9.17.0",
-    "@types/mocha": "^10.0.10",
-    "@types/node": "^22.10.2",
-    "@types/react": "^18.2.7",
-    "@types/react-dom": "^18.2.4",
-    "babel-preset-nano-react-app": "^0.1.0",
+    "@pandacss/dev": "^0.49.0",
+    "@types/react": "^19.0.2",
+    "@types/react-dom": "^19.0.2",
     "eslint": "^9.17.0",
-    "eslint-plugin-react": "^7.37.3",
+    "eslint-config-prettier": "^9.1.0",
+    "eslint-plugin-prettier": "^5.2.1",
+    "eslint-plugin-react": "^7.37.2",
     "globals": "^15.14.0",
-    "mocha": "^11.0.1",
-    "typescript": "^5.0.4",
-    "typescript-eslint": "^8.18.2",
-    "vite": "^4.5.5",
-    "vite-tsconfig-paths": "^4.2.0"
-  }
+    "postcss": "^8.4.49",
+    "prettier": "^3.4.2",
+    "typescript": "^5.7.2",
+    "typescript-eslint": "^8.18.1",
+    "vike": "^0.4.210",
+    "vite": "^6.0.5"
+  },
+  "type": "module"
 }
