@@ -1,6 +1,7 @@
-import { getBlogs } from "../src/utils/getBlogs";
+import { getBlogs } from "../utils/getBlogs";
 
-var assert = require("assert");
+import { strict as assert } from "assert";
+
 describe("Make sure we can get blog posts", function () {
   it("should return an array of blogs", function () {
     const blogs = getBlogs();
